requirements.txt
#
# These requirements were autogenerated by pipenv
# To regenerate from the project's Pipfile, run:
#
#    pipenv lock --requirements
#

-i https://pypi.org/simple
aiohttp==3.7.4.post0; python_version >= '3.6'
async-timeout==3.0.1; python_full_version >= '3.5.3'
attrs==21.2.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
chardet==4.0.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
click==8.0.3; python_version >= '3.6'
colorama==0.4.4; platform_system == 'Windows'
discord.py==1.7.3
flask==2.0.2
idna==3.3; python_version >= '3.5'
importlib-metadata==4.8.2; python_version < '3.8'
itsdangerous==2.0.1; python_version >= '3.6'
jinja2==3.0.3; python_version >= '3.6'
markupsafe==2.0.1; python_version >= '3.6'
multidict==5.2.0; python_version >= '3.6'
python-dotenv==0.19.2
typing-extensions==4.0.0; python_version < '3.8'
waitress==2.0.0
werkzeug==2.0.2; python_version >= '3.6'
yarl==1.7.2; python_version >= '3.6'
zipp==3.6.0; python_version >= '3.6'
